test: keep builtin catalog helper local
